requirements.txt
ultralytics
torch
torchvision
thop
psutil
pandas
pyyaml
